// apps/web/src/shared/api/generated/model/learningSubmissionSaveRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */
import type { LessonSubmissionSaveRequest } from './lessonSubmissionSaveRequest';
import type { ProblemSubmissionSaveRequest } from './problemSubmissionSaveRequest';

/**
 * 레슨 풀이 저장 request
 */
export interface LearningSubmissionSaveRequest {
  /** 레슨 풀이 제출 결과 */
  lessonSubmissionSaveRequest: LessonSubmissionSaveRequest;
  /** 문제 풀이 제출 리스트 */
  problemSubmissionSaveRequests: ProblemSubmissionSaveRequest[];
}