// apps/web/src/shared/api/generated/model/problemSubmissionSaveRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */

export interface ProblemSubmissionSaveRequest {
  /** 문제 아이디 */
  problemId: number;
  /** 최초 정/오답 여부 */
  isCorrect: boolean;
  /** 선택한 보기 아이디 (객관식) */
  selectedOptionId?: number;
  /** 제출한 답안 내용 (주관식) */
  submittedContent?: string;
}